requirements.txt
streamlit
python-dotenv
beautifulsoup4
requests
google-generativeai
telebot
pandas
